refactor:사이드바 상태저장 훅 shared로 이동
