+++ src/shared/model/index.ts
