+++ src/widgets/sidebar/ui/SideBar.tsx
@@ -2,7 +2,7 @@
 import { motion, AnimatePresence } from 'framer-motion';
 import SideBarInfo from '@/widgets/sidebar/ui/SideBarInfo';
 import SideBarNav from './SideBarNav';
-import { useDrawerStore } from './model';
+import { useDrawerStore } from '@/shared/model';
 import { cn } from '@/shared/utils/cn';
 
 export default function SideBar() {
